Create DTXChart module with 2 classes
Parser takes in a .dtx or .gda file, creates and store a dtxData object.
LinePositionMapper parse a raw dtxData object, computes and store starting bpm and absolute positions of bar start line & bpm markers for each and every bar.
It also provides a method to get absolute position of any line given the bar-number and line-number.