Added a barAligned drawing mode
Updated Charter to draw each page with full bars only. The bottomline of each page always starts with at a bar.

Draw the height of each page according to the number of bars it can fit. This would result in uneven page height for songs that change BPM or time signature within the song